+++ Orbuculum/Changelog.md
@@ -1,5 +1,13 @@
 ﻿# Changelog
 
+## Version 1.0.3.0
+- Added `🧙 Auto Balancing Exposure` - An instruction to automatically balance exposure count by progress and weight with a defined list of exposure info
+- Added `🕑 Loop While Hour Angle` and `🕑 Loop While Next Target Hour Angle` loop conditions to loop using hour angle
+- Fixed some issues where the existing conditions did not copy all relevant settings
+
+## Version 1.0.2.0
+- Move code to separate repository
+
 ## Version 1.0.1.0
 
 - Added offset for `🔮 Loop While Next Target Is Below Horizon` instruction